dt-binding: Add Tegra186/Tegra194 BPMP I2C binding
In Tegra186/Tegra194, the BPMP (Boot and Power Management Processor) owns
certain HW devices, such as the I2C controller for the power management I2C
bus. Software running on other CPUs must perform IPC to the BPMP in order to
execute transactions on that I2C bus. This binding describes an I2C bus
that is accessed in such a fashion

Change-Id: If24917413cfe94481b9a18d4d7fdaa370df9d340
Signed-off-by: Rajesh Gumasta <rgumasta@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2287011
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Shardar Mohammed <smohammed@nvidia.com>
Reviewed-by: Suresh Venkatachalam <skathirampat@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>