drivers: staging: Add Generic timestamping support
Adds generic hardware timestamping (GTE) support for the T194 SoC.
There are certain applications like robotics which may require more
accurate recording of the occurance of the certain events. GTE driver
will help monitor AON GPIOs and LIC interrupts.

GTE driver also implements char driver for userspace to monitor GPIO
through IOCTL.

Other kernel APIs are exported and defined in the tegra-gte.h file for
kernel client drivers to use. APIs are not stable and subject to change
in the future. It is for this reason GTE driver is provided in the
staging directory.

Since its APIs are not stable, it should not break existing nvpps driver
code from the stage-main which also uses GTE but in limited way. It is
for this reason this driver will not be selected if nvpps is enabled.

Bug 2757864

Change-Id: I0947f2b90232eb6c2a31163e33ec5ad45b7bd415
Signed-off-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2287452
(cherry picked from commit 6ff9cac2cad40cf6fd054410f1f965feb61f5f92 in rel-32)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2308778
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit