firmware: tegra: fix channel reclaim feature
Commit 4710a191d45f9889d204bc06d40ee4cec077eef4 (firmware: tegra:
reclaim timedout channels) introduced a feature to reclaim channels
that timed out. The implementation has a logical bug that upon
receiving a message, and if there was any channel to be reclaimed
(regardless of the current channel), the channel was freed forcefully.

Fix the issue by correcting the boolean statement to use bit-wise-and
of bit-masks.

Bug 200588424

Change-Id: Ifcca42e1d1a8a9ef534af1e421d8340e93eea878
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2299496
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>