t19x: galen: Add overlay for AGX Xavier
Add the following device-tree overlay for the Jetson AGX Xavier platform.

1. Respeaker 4 Mic Array Card

The overlays for the ReSpeaker 4 Mic Voicecard add-on module is developed
to be directly applied to the base DTB file for the Jetson platform.
For example, these overlays can be used as follows in order to generate
a complete DTB with the overlay applied.

$ fdtoverlay -i <base-dtb> -o <out-dtb> <overlay-dtbo>

Bug 200526593

Change-Id: I73568ff5854fbc37c5338ad2e32d3fc5bafe7342
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/stardust-dts/+/2304081
(cherry picked from commit b33da44acc5d865921a137485d07cf02c6171100)
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/stardust-dts/+/2321792
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Asha Talambedu <atalambedu@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Asha Talambedu <atalambedu@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>