t19x: jakku: Add overlay for ReSpeaker 4Mic
Add the following device-tree overlays for the Xavier NX platform.

1. Respeaker 4 Mic Array Card

The overlays for the ReSpeaker 4 Mic Voicecard add-on module is developed
to be directly applied to the base DTB file for the Jetson platform.
For example, these overlays can be used as follows in order to generate
a complete DTB with the overlay applied.

$ fdtoverlay -i <base-dtb> -o <out-dtb> <overlay-dtbo>

Bug 200526593

Change-Id: Ia39355e3b1f5999affcfd41fe15e28ab7a719337
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/jakku-dts/+/2304094
(cherry picked from commit 67cefdac423de5c59efee534a9b162a2136d9895)
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/jakku-dts/+/2321799
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Asha Talambedu <atalambedu@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Asha Talambedu <atalambedu@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit