nvidia: sound: exclude 'kernel/nvidia' drivers
5.4 kernel for HDA uses the upstreamed hda_tegra.c driver and pending
changes are being cherry-picked to 5.4 kernel.

For AHUB drivers the upstream is under progress and we plan to use
5.4 kernel for these drivers as well. Already the drivers are copied
for the new kernel.

Currently in menuconfig SND_HDA_TEGRA appears twice and it shows the
value as 'm'. For *_TEGRA_ALT configuration, though the configs are
in disabled state, these options are available. This leads to confusion.
This patch thus excludes "kernel/nvidia" drivers for HDA and AHUB on
5.4 kernel to make it look more cleaner.

Bug 2845498

Change-Id: Iaeb0d9cfd521573acf75114967ab76ed4b0016fe
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2307961
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit