dts: soc: t19x: Add Nvlink power domain node
Nvlink power domain was made always on in BPMP DT. But BPMP-server
sees that the power domain is in OFF state when queried through MRQ.
This is because the state returned by MRQ query is not _physical_
state of the partition, but the _requested_ state, i.e. the state as
it set by MRQ requests. Hence Guest OS need to send unpowergate
request even though it is already unpowergated by BPMP-FW. This
change adds Nvlink power domain node in DT for the same.

Bug 200432470

Change-Id: I2f0f9d9fb72044f083c19b41e6932ffab700128f
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t19x/+/2317191
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>