thermal: Kernel driver for OC event handling
Why?
In case of an OC(Over Current) event in Tegra SOC,
BPMP gets an interrupt indicating the same. However,
there is no indication to ccplex and userspace about
the same. This driver provides the interface between
BPMP OC implementation and userspace.

How?
Once there is an EDP OC interrupt to BPMP, BPMP writes
to a Shared mailbox register dedicated for BPMP and CCPLEX
communication for EDP OC. Once the interrupt is received,
Kernel reads the OC status registers to realize the state
of the EDP OC HW. The kernel driver provides interfaces
for Userspace to poll for the status/interrupts count.

Bug 1688327

Change-Id: Iec57849b2be8343bc9d8c617bffd44ab63c8fa9d
Signed-off-by: Mantravadi Karthik <mkarthik@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2328163
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit