serial: tegra: avoid null pointer access
UART isr could already be running for Rx after HW interrupt is
disabled from UART shutdown sequence. If it's running, the HW
interrupts could be re-enabled by the isr. And it may access
to the freed variables and lead to null pointer access

This fix is to disable irq prior to HW interrupt. And it adds
check for the variables

Bug 2415388

Change-Id: I7b4d8d023f1001a0000e382241fae6b26e3c0777
Signed-off-by: Ahung Cheng <ahcheng@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1928067
Reviewed-on: https://git-master.nvidia.com/r/1974241
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.4/+/2321665
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shardar Mohammed <smohammed@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>