dmaengine: tegra210-adma: override ADMA FIFO size
ADMAIF FIFO uses a ring buffer and it is divided amongst the available
channels. The default FIFO size of ADMAIF TX/RX channels is as below:
 * On Tegra210,
     channel 1 to 2 : size = 0x3
     channel 3 to 10: size = 0x2
 * On Tegra186 and later,
     channel 1 to 4 : size = 0x3
     channel 5 to 20: size = 0x2

As per recommendation from HW, FIFO size of ADMA channel should be
same as the corresponding ADMAIF channel it maps to. FIFO corruption
was observed if the sizes do not match. We are using default FIFO
sizes on ADMAIF side and till now never got requests to have different
values than the default ones.

Thus the index of the channel is maintained after which the default
size changes and this member is referred as 'slave_id'. As per the
current ADMAIF channel index the size is overridden for the mapped
ADMA channel.

This is partially cherry-picked from following commits:
 commit 1b490e4f1e74 ("dma: tegra: Disable ADMA WAR from T194 chip")
 commit fe0a79398750 ("dma: tegra: support for t186 adma")
 commit 2d58b550a38d ("dmaengine: tegra210-adma: simplify chip_data")
 commit 89e21db63ddb ("dma: tegra: unification changes for t210 driver")

Original corruption issue was raised in bug 200220715.

Bug 2845498

Change-Id: I35f756258c573be38599849b857f751c36156043
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.4/+/2328509
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Dipesh Gandhi <dipeshg@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>