UPSTREAM: soc/tegra: fuse: Update the SoC revision attribute to display a name
Currently the SoC revision attribute for Tegra devices displays the
value of the enum associated with a particular revision. This is not
very useful because to obtain the actual revision you need to
use the tegra_revision enumeration to translate the value.

It is more meaningful to display a name for the revision, such as
'A01', than the enumarated value and therefore, update the revision
attribute to display a name. This change does alter the ABI, which
is unfortunate, but this is more meaningful and maintable.

Bug 200551105

Change-Id: I0f2be1a5cf033f3cec2c4c581a8042eb9c52c07c
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2332597
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
(cherry picked from commit d4a2d55a5f349105484679c523f215b021493c07)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.4/+/2335853
Tested-by: Bitan Biswas <bbiswas@nvidia.com>