Revert "platform: t19x: common: add k5.4 ethernet overrides"
This reverts commit 36690a2f7aa2c87f6202ba854cdf6079c070b323.

Reason to revert: This change was made to load the upstream
driver stmmac using the downstream DT since we support only
downstream DT. The upstream driver was needed to support gvs
and the sanity on TOT for ethernet because nvethernet driver
was not merged.

Now with this change nvethernet driver is also getting enabled,
so we do not need to load stmmac driver for ethernet anymore.

Hence removing this Change

Bug 200591808

Signed-off-by: Bibhay Ranjan <bibhayr@nvidia.com>
Change-Id: I77810d82c93274664fdf31133697f2b7696115c6
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/common/+/2337273
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>