kconfig: revert to only use /dev/null
An upstream change to deal with some assembler not being able to
use /dev/null as input and output led to a change introducing the
use of /dev/zero. This is causing clean up issues with the internal
unix-build tool so reverting to only using /dev/null since the
assembler we use does not have any issues with /dev/null being passed
twice in args.

Bug 2967443
Bug 200617764

Signed-off-by: Nicolas Benech <nbenech@nvidia.com>