[TEMP] t19x: common: disable audio on super-module
On 5.4 TOT, sound card registration fails because audio codecs, present
on super-module, are not getting probed. This happens because the I2C
master routed to 40-pin is not getting probed and has dependency on
host1x driver bringup.

This patch temporarily disables audio codecs on super-module and sound
card is registered with only on-board codec. Once required dependencies
are met, this patch can be reverted.

Bug 200611040
Bug 200610798

Change-Id: I57f4182d186e8f9a993b2f4f5b3c9d3e6d434b91
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/common/+/2340567
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit