[TEMP] t19x: galen: disable super-module audio
On 5.4 TOT, sound card registration fails because audio codecs, present
on super-module, are not getting probed. This happens because the I2C
master routed to 40-pin is not getting probed and has dependency on
host1x driver bringup.

This patch temporarily disables audio codecs on super-module and sound
card is registered with only on-board codec. Once required dependencies
are met, this patch can be reverted.

Bug 200611040
Bug 200610798

Change-Id: Idbe711c2e57b417f6bc91fbfbe45a5d2eac4b58f
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/stardust-dts/+/2341233
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit