DOWNSTREAM: arm64: dts: Add support for booting with ACPI on Xavier
Add a device-tree variant for booting Linux with ACPI on Jetson Xavier.
The main different between the device-tree variant used when booting
with ACPI is the serial port configuration. By default the TCU uses
UARTC on Xavier, but when booting with ACPI the TCU is not supported.
It is possible to flash Jetson Xavier to disable the TCU and use the
8250 driver on UARTC, but this means that it is necessary to re-flash
Xavier everytime you switch booting Linux with device-tree and ACPI.

To simplify matters, enable UARTB for booting Linux with ACPI and
disable the TCU. Note that it is also possible to boot Linux with
device-tree and use UARTB for the serial console to avoid re-flashing
the platform.

Bug 2877970

Change-Id: Id475959b0b7ee63fa83c9e527b2ac93c3d4fe642
Signed-off-by: Ashish Singhal <ashishsingha@nvidia.com>
Signed-off-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.4/+/2343068
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit